исправление  страницы Билеты по FSD
