+++ src/entities/ticket/index.ts
@@ -1,0 +1,2 @@
+export type { Ticket } from './model/types';
+export { ticketsMockData } from './model/mock';
