+++ src/entities/ticket/model/types.ts
