+++ src/pages/ticketsPage/index.ts
@@ -1,0 +1,1 @@
+export { TicketsPage } from './ui/ticketsPage.tsx';
