--- src/pages/ticketsPage/index.tsx
@@ -1,1 +1,0 @@
-export { TicketsPage } from './ticketsPage.tsx';
