+++ src/shared/lib/date/formatters.ts
@@ -1,0 +1,5 @@
+import moment from 'moment';
+
+export const formatDateTime = (dateTime: string): string => {
+  return moment(dateTime).format('YYYY-MM-DD, HH:mm');
+};
