+++ src/shared/lib/date/index.ts
@@ -1,0 +1,1 @@
+export { formatDateTime } from './formatters';
