+++ src/widgets/layouts/Layout.tsx
@@ -1,16 +1,14 @@
-import type { FC, ReactNode } from 'react';
-
-import { Header } from '@widgets/header';
 import { Footer } from '@widgets/footer';
+import { Header } from '@widgets/header';
 
-interface LayoutProps {
-  children: ReactNode;
-}
+import { Outlet } from 'react-router-dom';
 
-export const Layout: FC<LayoutProps> = ({ children }) => (
+export const Layout = () => (
   <div className="flex flex-col min-h-screen">
     <Header />
-    <main className="flex-grow">{children}</main>
+    <main className="flex-grow">
+      <Outlet />
+    </main>
     <Footer />
   </div>
 );
